chore: Restructure App Utilities.
