+++ src/ansys/fluent/core/services/app_utilities.py
@@ -30,7 +30,6 @@ import grpc
 
 from ansys.api.fluent.v0 import app_utilities_pb2 as AppUtilitiesProtoModule
 from ansys.api.fluent.v0 import app_utilities_pb2_grpc as AppUtilitiesGrpcModule
-from ansys.fluent.core._types import PathType
 from ansys.fluent.core.services._protocols import ServiceProtocol
 from ansys.fluent.core.services.interceptors import (
     BatchInterceptor,
@@ -38,10 +37,9 @@ from ansys.fluent.core.services.interceptors import (
     GrpcErrorInterceptor,
     TracingInterceptor,
 )
-from ansys.fluent.core.streaming_services.events_streaming import SolverEvent
 
 
-class AppUtilitiesService(ServiceProtocol):
+class AppUtilitiesServiceV0(ServiceProtocol):
     """AppUtilities Service."""
 
     def __init__(  # pyright: ignore[reportMissingSuperCall]
@@ -163,341 +161,16 @@ class AppUtilitiesService(ServiceProtocol):
         return self._stub.SetWorkingDirectory(request, metadata=self._metadata)
 
 
-@dataclass
-class ProcessInfo:
-    """ProcessInfo dataclass to hold process information."""
+class AppUtilitiesService(AppUtilitiesServiceV0):
+    """AppUtilities Service (v1 proto API)."""
 
-    process_id: int
-    hostname: str
-    working_directory: str
+    def _create_stub(self, intercept_channel):
+        """Create the v1 gRPC stubs.
 
-
-@dataclass
-class BuildInfo:
-    """BuildInfo dataclass to hold build information."""
-
-    build_time: str
-    build_id: str
-    vcs_revision: str
-    vcs_branch: str
-
-
-class AppUtilitiesOld:
-    """AppUtilitiesOld."""
-
-    def __init__(self, scheme_eval):
-        """__init__ method of AppUtilitiesOld class."""
-        self.scheme = scheme_eval
-
-    def get_product_version(self) -> str:
-        """Get product version."""
-        return self.scheme.version
-
-    def get_build_info(self) -> dict:
-        """Get build info."""
-        build_time = self.scheme.eval("(inquire-build-time)")
-        build_id = self.scheme.eval("(inquire-build-id)")
-        vcs_revision = self.scheme.eval("(inquire-src-vcs-id)")
-        vcs_branch = self.scheme.eval("(inquire-src-vcs-branch)")
-        return BuildInfo(
-            build_time=build_time,
-            build_id=build_id,
-            vcs_revision=vcs_revision,
-            vcs_branch=vcs_branch,
-        )
-
-    def get_controller_process_info(self) -> dict:
-        """Get controller process info."""
-        cortex_host = self.scheme.eval("(cx-cortex-host)")
-        cortex_pid = self.scheme.eval("(cx-cortex-id)")
-        cortex_pwd = self.scheme.eval("(cortex-pwd)")
-        return ProcessInfo(
-            process_id=cortex_pid,
-            hostname=cortex_host,
-            working_directory=cortex_pwd,
-        )
-
-    def get_solver_process_info(self) -> dict:
-        """Get solver process info."""
-        fluent_host = self.scheme.eval("(cx-client-host)")
-        fluent_pid = self.scheme.eval("(cx-client-id)")
-        fluent_pwd = self.scheme.eval("(cx-send '(cx-client-pwd))")
-        return ProcessInfo(
-            process_id=fluent_pid,
-            hostname=fluent_host,
-            working_directory=fluent_pwd,
-        )
-
-    def get_app_mode(self) -> Enum:
-        """Get app mode."""
-        from ansys.fluent.core import FluentMode
-
-        if self.scheme.eval("(cx-solver-mode?)"):
-            mode_str = self.scheme.eval('(getenv "PRJAPP_APP")')
-            if mode_str == "flaero_server":
-                return FluentMode.SOLVER_AERO
-            elif mode_str == "flicing":
-                return FluentMode.SOLVER_ICING
-            else:
-                return FluentMode.SOLVER
-        else:
-            return FluentMode.MESHING
-
-    def start_python_journal(self, journal_name: str | None = None) -> int:
-        """Start python journal."""
-        if journal_name:
-            self.scheme.exec([f'(api-start-python-journal "{journal_name}")'])
-        else:
-            self.scheme.eval("(define pyfluent-journal-str-port (open-output-string))")
-            self.scheme.eval("(api-echo-python-port pyfluent-journal-str-port)")
-            return "1"
-
-    def stop_python_journal(self, journal_id: str | None = None) -> str:
-        """Stop python journal."""
-        if journal_id:
-            self.scheme.eval("(api-unecho-python-port pyfluent-journal-str-port)")
-            journal_str = self.scheme.eval(
-                "(close-output-port pyfluent-journal-str-port)"
-            )
-            return journal_str
-        else:
-            self.scheme.exec(["(api-stop-python-journal)"])
-
-    def is_beta_enabled(self) -> bool:
-        """Is beta enabled."""
-        return self.scheme.eval("(is-beta-feature-available?)")
-
-    def enable_beta(self):
-        """Enable beta features.
-
-        Raises
-        ------
-        RuntimeError
-            Not supported before Fluent 2025 R2.
+        In addition to the main ApplicationRuntime stub, an Events stub is
+        created here because the pause/resume RPCs migrated to the Events
+        service in the v1 proto API.  A FieldData stub and a Settings stub are
+        also created because IsSolutionDataAvailable and IsWildcard migrated to
+        those services respectively.
         """
-        raise RuntimeError(
-            "Enabling beta is not supported by PyFluent for Fluent versions before 2025 R2."
-        )
-
-    def is_wildcard(self, input: str | None = None) -> bool:
-        """Is wildcard."""
-        return self.scheme.eval(f'(has-fnmatch-wild-card? "{input}")')
-
-    def is_solution_data_available(self) -> bool:
-        """Is solution data available."""
-        return self.scheme.eval("(data-valid?)")
-
-    def register_pause_on_solution_events(self, solution_event: SolverEvent) -> int:
-        """Register pause on solution events."""
-        unique_id: int = self.scheme.eval(
-            f"""
-            (let
-                ((ids
-                    (let loop ((i 1))
-                        (define next-id (string->symbol (format #f "pyfluent-~d" i)))
-                        (if (check-monitor-existence next-id)
-                            (loop (1+ i))
-                            (list i next-id)
-                            )
-                        )
-                    ))
-                (register-solution-monitor
-                    (cadr ids)
-                    (lambda (niter time)
-                        (if (integer? niter)
-                            (begin
-                                (events/transmit 'auto-pause (cons (car ids) niter))
-                                (grpcserver/auto-pause (is-server-running?) (cadr ids))
-                                )
-                            )
-                        ()
-                        )
-                    {"#t" if solution_event == SolverEvent.TIMESTEP_ENDED else "#f"}
-                    )
-                (car ids)
-                )
-        """
-        )
-        return unique_id
-
-    def resume_on_solution_event(self, registration_id: int) -> None:
-        """Resume on solution event."""
-        self.scheme.eval(
-            f"(grpcserver/auto-resume (is-server-running?) 'pyfluent-{registration_id})"
-        )
-
-    def unregister_pause_on_solution_events(self, registration_id: int) -> None:
-        """Unregister pause on solution events."""
-        self.scheme.eval(f"(cancel-solution-monitor 'pyfluent-{registration_id})")
-
-    def exit(self) -> None:
-        """Exit."""
-        self.scheme.exec(("(exit-server)",))
-
-    def set_working_directory(self, path: PathType) -> None:
-        """Change client cortex dir."""
-        self.scheme.eval(f'(syncdir "{os.fspath(path)}")')
-
-
-class AppUtilities:
-    """AppUtilities."""
-
-    def __init__(self, service: AppUtilitiesService):
-        """__init__ method of AppUtilities class."""
-        self.service = service
-
-    def get_product_version(self) -> str:
-        """Get product version."""
-        request = AppUtilitiesProtoModule.GetProductVersionRequest()
-        response = self.service.get_product_version(request)
-        return f"{response.major}.{response.minor}.{response.patch}"
-
-    def get_build_info(self) -> dict:
-        """Get build info."""
-        request = AppUtilitiesProtoModule.GetBuildInfoRequest()
-        response = self.service.get_build_info(request)
-        return BuildInfo(
-            build_time=response.build_time,
-            build_id=response.build_id,
-            vcs_revision=response.vcs_revision,
-            vcs_branch=response.vcs_branch,
-        )
-
-    def get_controller_process_info(self) -> dict:
-        """Get controller process info."""
-        request = AppUtilitiesProtoModule.GetControllerProcessInfoRequest()
-        response = self.service.get_controller_process_info(request)
-        return ProcessInfo(
-            process_id=response.process_id,
-            hostname=response.hostname,
-            working_directory=response.working_directory,
-        )
-
-    def get_solver_process_info(self) -> dict:
-        """Get solver process info."""
-        request = AppUtilitiesProtoModule.GetSolverProcessInfoRequest()
-        response = self.service.get_solver_process_info(request)
-        return ProcessInfo(
-            process_id=response.process_id,
-            hostname=response.hostname,
-            working_directory=response.working_directory,
-        )
-
-    def get_app_mode(self) -> Enum:
-        """Get app mode.
-
-        Raises
-        ------
-        ValueError
-            If app mode is unknown.
-        """
-        import ansys.fluent.core as pyfluent
-
-        request = AppUtilitiesProtoModule.GetAppModeRequest()
-        response = self.service.get_app_mode(request)
-        match response.app_mode:
-            case AppUtilitiesProtoModule.APP_MODE_UNKNOWN:
-                raise ValueError("Unknown app mode.")
-            case AppUtilitiesProtoModule.APP_MODE_MESHING:
-                return pyfluent.FluentMode.MESHING
-            case AppUtilitiesProtoModule.APP_MODE_SOLVER:
-                return pyfluent.FluentMode.SOLVER
-            case AppUtilitiesProtoModule.APP_MODE_SOLVER_ICING:
-                return pyfluent.FluentMode.SOLVER_ICING
-            case AppUtilitiesProtoModule.APP_MODE_SOLVER_AERO:
-                return pyfluent.FluentMode.SOLVER_AERO
-
-    def start_python_journal(self, journal_name: str | None = None) -> int:
-        """Start python journal."""
-        request = AppUtilitiesProtoModule.StartPythonJournalRequest()
-        if journal_name:
-            request.journal_name = journal_name
-        response = self.service.start_python_journal(request)
-        return response.journal_id
-
-    def stop_python_journal(self, journal_id: str | None = None) -> str:
-        """Stop python journal."""
-        request = AppUtilitiesProtoModule.StopPythonJournalRequest()
-        if journal_id:
-            request.journal_id = journal_id
-        response = self.service.stop_python_journal(request)
-        return response.journal_str
-
-    def is_beta_enabled(self) -> bool:
-        """Is beta enabled."""
-        request = AppUtilitiesProtoModule.IsBetaEnabledRequest()
-        response = self.service.is_beta_enabled(request)
-        return response.is_beta_enabled
-
-    def enable_beta(self) -> None:
-        """Enable beta features."""
-        request = AppUtilitiesProtoModule.EnableBetaRequest()
-        self.service.enable_beta(request)
-
-    def is_wildcard(self, input: str | None = None) -> bool:
-        """Is wildcard."""
-        request = AppUtilitiesProtoModule.IsWildcardRequest()
-        request.input = input
-        response = self.service.is_wildcard(request)
-        return response.is_wildcard
-
-    def is_solution_data_available(self) -> bool:
-        """Is solution data available."""
-        request = AppUtilitiesProtoModule.IsSolutionDataAvailableRequest()
-        response = self.service.is_solution_data_available(request)
-        return response.is_solution_data_available
-
-    def register_pause_on_solution_events(self, solution_event: SolverEvent) -> int:
-        """Register pause on solution events."""
-        request = AppUtilitiesProtoModule.RegisterPauseOnSolutionEventsRequest()
-        request.solution_event = AppUtilitiesProtoModule.SOLUTION_EVENT_UNKNOWN
-        match solution_event:
-            case SolverEvent.ITERATION_ENDED:
-                request.solution_event = (
-                    AppUtilitiesProtoModule.SOLUTION_EVENT_ITERATION
-                )
-            case SolverEvent.TIMESTEP_ENDED:
-                request.solution_event = (
-                    AppUtilitiesProtoModule.SOLUTION_EVENT_TIME_STEP
-                )
-        response = self.service.register_pause_on_solution_events(request)
-        return response.registration_id
-
-    def resume_on_solution_event(self, registration_id: int) -> None:
-        """Resume on solution event."""
-        request = AppUtilitiesProtoModule.ResumeOnSolutionEventRequest()
-        request.registration_id = registration_id
-        self.service.resume_on_solution_event(request)
-
-    def unregister_pause_on_solution_events(self, registration_id: int) -> None:
-        """Unregister pause on solution events."""
-        request = AppUtilitiesProtoModule.UnregisterPauseOnSolutionEventsRequest()
-        request.registration_id = registration_id
-        self.service.unregister_pause_on_solution_events(request)
-
-    def exit(self) -> None:
-        """Exit."""
-        request = AppUtilitiesProtoModule.ExitRequest()
-        self.service.exit(request)
-
-    def set_working_directory(self, path: PathType) -> None:
-        """Change client cortex dir."""
-        request = AppUtilitiesProtoModule.SetWorkingDirectoryRequest()
-        request.path = os.fspath(path)
-        self.service.set_working_directory(request)
-
-
-class AppUtilitiesV252(AppUtilities):
-    """AppUtilitiesV252.
-    This is for methods whose implementations are missing in the 25R2 server.
-    """
-
-    def __init__(self, service: AppUtilitiesService, scheme):
-        super().__init__(service)
-        self.scheme = scheme
-
-    def enable_beta(self) -> None:
-        """Enable beta features."""
-        self.scheme.eval(
-            '(fl-execute-cmd "file" "beta-settings" (list (cons "enable?" #t)))'
-        )
+        return AppUtilitiesGrpcModule.ApplicationRuntimeStub(intercept_channel)
